requirements.txt
streamlit==1.49.1
tensorflow-cpu==2.15.1
numpy==1.26.4
pillow>=9.5
matplotlib>=3.7,<3.9
gTTS>=2.5
h5py>=3.8,<4
